wallet function added,check dates func added,manager transfering added
we now check if the event date has passed or not. if it is passed users cant interact with that event anymore. we added wallet function. user can add money to wallet and buy tickets using the wallet money.  also added if eventholder account is deleted event manager  is transferred to admin.